+++ src/app/admin/rounds/[roundId]/matchDays/[matchDayId]/match/new/page.tsx
@@ -1,4 +1,4 @@
-import { MatchForm } from "@/components/result-form/result-form";
+import { MatchForm } from "@/components/admin/rounds/match-days/match/match-form";
 import { getTeams } from "@/lib/actions/teams";
 import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
 import Link from "next/link";
