+++ src/app/admin/rounds/[roundId]/matchDays/new/page.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
-import MatchDayForm from "@/components/rounds/matchDays/matchDayForm";
+import MatchDayForm from "@/components/admin/rounds/match-days/match-day-form";
 
 const NewMatchDayPage = async ({ params }: Params) => {
   return (
