+++ src/app/admin/teams/[teamId]/edit/page.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
 import { getTeam } from "@/lib/actions/teams";
-import TeamForm from "@/components/teams/teamForm";
+import TeamForm from "@/components/admin/teams/teamForm";
 
 const EditTeamPage = async ({ params }: Params) => {
   const team = await getTeam(params.teamId);
