+++ src/components/admin/rounds/match-days/match/match-form.tsx
@@ -6,7 +6,7 @@ import { ITeam } from "@/lib/models/team";
 import { useRouter } from "next/navigation";
 import { useEffect } from "react";
 import { useFormState } from "react-dom";
-import Form from "../form/form";
+import Form from "../../../../form/form";
 
 interface FormProps {
   match?: IMatch;
