+++ src/components/main/top-scorers/top-scorers.tsx
